virt-controller-vmi: separate lifecycle related functions
This moves all lifecycle related functions and event handlers into lifecycle.go
